requirements.txt
flask==3.0.2
flask-cors==4.0.0
icalendar==5.0.11
python-dateutil==2.8.2
gunicorn==21.2.0
pytest==8.0.2
